Move permissions handling into a QueryHook
This moves permission handling into a new QueryHook implementation
called `PermissionsHook`

The `AuthManager` is now moved from `HandlerFactory` and
`DfSessionService` and is now held only inside the `PermissionsHook`
struct.

This does not yet move from string processing to query matching, nor do
we yet try to completely decouple the permissions logic completely from
everything else.
