+++ datafusion-postgres/src/hooks/mod.rs
@@ -1,3 +1,4 @@
+pub mod permissions;
 pub mod set_show;
 pub mod transactions;
